+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "mcm-js-commons",
-  "version": "0.0.26",
+  "version": "0.0.29",
   "private": false,
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
