+++ src/units/link/Link/index.tsx
@@ -1,13 +1,12 @@
 import Link from "next/link";
 import styled from "@emotion/styled";
-import CommonsHooksComponents from "../../../hooks";
+import { getAllComponentsClassName } from "../../../hooks";
 
 import { LinkUnitTypes } from "../../../types/units";
 import _Error from "../../error";
 
 // router 이동 관련 컴포넌트
 export default function _Link(props: LinkUnitTypes) {
-  const { getAllComponentsClassName } = CommonsHooksComponents();
   const { href, children, className, styles } = props;
 
   return (
