+++ src/units/text/span/index.tsx
@@ -1,11 +1,10 @@
 import { SpanText } from "..";
 
-import CommonsHooksComponents from "../../../hooks";
+import { getAllComponentsClassName } from "../../../hooks";
 import { TextUnitTypes } from "../../../types/units";
 
 // span 태그를 출력하는 컴포넌트
 export default function _SpanText(props: TextUnitTypes) {
-  const { getAllComponentsClassName } = CommonsHooksComponents();
   const { styles, className, children, isError } = props;
 
   return (
